gcc-4.8: On OSX, pre-unlink.sh is a complete no-op.

+++ gcc-4.8/pre-unlink.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ "$(uname)" == "Darwin" ]; then
+    # The post-link script only runs on Linux,
+    # so this pre-unlink script isn't needed on OSX.
+    exit 0;
+fi
+
 # Remove the crt symlinks created in post-link.sh
 find "$PREFIX"/lib/gcc/*/* -type l -print0 | xargs -0 rm -f
